Use proper tag for mocha support in standard style tester
Initially I stopped the standard style tester from getting upset about
'describe' and 'it' by using the global tag in a comment but found that
it has a more descriptive way built in.

This commit sets the build environment of the standard style checker to
mocha inside the test file.