Refactor underscore prefixes and others
The underscores were useless and to be honest I can't quite remember why
there were so many.

During this I renamed _all to droneList which is a more descriptive name
and doesn't have a stupid underscore.

I also changed the Drones.all() function to be a getter method so
Drones.all is the new standard.